fix: C51 变参 printf uint8_t 粘连错位 + 时间设置页布局重排
根因：Keil C51 变参函数中 char/uint8_t 参数按 1 字节压栈，
而 %d 按 2 字节读取 → 相邻字段粘连成 16 位值：
  月份 "2066" = [month][day] 0x0812，分钟 "1024" = [minute][second]
  0x0400，秒 "2060" = [second][下一位]（调整幅度 256 同源）。
修复：所有变参 printf/CalcStringWidth 的 uint8_t 参数显式 (int) 强转：
  - 时间设置页日期/时间行（TimeSetAuxFunc）
  - 万年历页日期行（ClockAuxFunc，wpx 计算 + 绘制）
  - 主菜单顶部时钟（OLED_UI_ShowClock）——此前 14:00:00 时相邻字节
    恰为 0 而"凑巧正常"，改默认 00:00:00 后秒数一走即暴露

时间设置页布局重排（用户提议）：
  - 删除底部按键说明行
  - 两行统一 12x12 汉字 + 7x12 数字字号，水平居中
  - 反色选择框按新布局精确对齐字段（不再偏移）

+++ src/ui/OLED_UI/OLED_UI.c
@@ -110,7 +110,7 @@ void OLED_UI_ShowClock(void)
         last_h = t.hour;
         last_m = t.minute;
         last_s = t.second;
-        OLED_Printf(0, 0, OLED_6X8_HALF, "%02d:%02d:%02d", t.hour, t.minute, t.second);
+        OLED_Printf(0, 0, OLED_6X8_HALF, "%02d:%02d:%02d", (int)t.hour, (int)t.minute, (int)t.second);   /* uint8_t 需 (int) 强转，防变参粘连 */
     }
 }
 /**
